Added show-areas-functions to 'ShowDishForm' and changed weight type from REAL to INTEGER

+++ src/adddishform.cpp
@@ -104,7 +104,7 @@ AddDishForm::AddDishForm(QSqlTableModel *categories_model, QWidget *parent) : QD
     connect(picture_select_btn, SIGNAL(clicked()), this, SLOT(select_image()));
 
     connect(accept_btn, SIGNAL(clicked()), this, SLOT(add_dish()));
-    connect(this, SIGNAL(add_dish(QString,double,double,QString,int,QString)), parent, SLOT(add_dish(QString,double,double,QString,int,QString)));
+    connect(this, SIGNAL(add_dish(QString,int,double,QString,int,QString)), parent, SLOT(add_dish(QString,int,double,QString,int,QString)));
 }
 
 void AddDishForm::select_image()
