+++ src/adddishform.h
@@ -51,7 +51,7 @@ public:
     bool loading_image();
 
 signals:
-    void add_dish(QString name, double weight, double price, QString category, int estimated_time, QString url);
+    void add_dish(QString name, int weight, double price, QString category, int estimated_time, QString url);
 
 public slots:
     void select_image();
